backend_app/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
bcrypt==4.2.1
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
Mako==1.3.6
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2==2.9.10
pyasn1==0.6.1
pydantic==2.10.2
pydantic_core==2.27.1
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
